+++ pkg/common/hs_types.go
@@ -46,8 +46,6 @@ type HSVolume struct {
 	Path                   string
 	VolumeMode             string
 	SourceSnapPath         string
-	SourceSnapVolumePath   string
-	SourceSnapFilePaths    []string
 	FSType                 string
 	Comment                string
 	SourceSnapShareName    string
@@ -172,6 +170,7 @@ type File struct {
 	Name     string         `json:"name"`
 	Path     string         `json:"path"`
 	Size     int64          `json:"size"`
+	Type     string         `json:"type"`
 	Children []FileChildren `json:"children"`
 }
 
@@ -179,6 +178,7 @@ type FileChildren struct {
 	Name       string `json:"name"`
 	Path       string `json:"path"`
 	Size       int64  `json:"size"`
+	Type       string `json:"type"`
 	Parent     string `json:"parent"`
 	SharePath  string `json:"sharePath"`
 	ShareName  string `json:"shareName"`